rename workspace_conditions to workspace
