+++ src/lib.rs
@@ -4,4 +4,4 @@ pub mod layout;
 pub mod matches;
 pub mod picker;
 pub mod tmux;
-pub mod workspace_conditions;
+pub mod workspace;
